+++ Mind/CommonFunction.cpp
@@ -50,7 +50,7 @@ namespace Mind
 			shared_ptr<iDataArray> res(new DataArray(conceptSet->BaseConceptCount()));
 
 			vector<shared_ptr<iConcept>> conceptSequence=chain->GetConceptVec();
-			for (unsigned int i=0;i<conceptSequence.size();++i)
+			for (size_t i=0;i<conceptSequence.size();++i)
 			{
 				shared_ptr<BaseConcept> base=dynamic_pointer_cast<BaseConcept>(conceptSequence[i]);
 				if(base==NULL)
